modified project to function as a webapi
