+++ SeaSkyNavigator.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34525.116
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SeaSkyNavigator", "WeatherApp\SeaSkyNavigator.csproj", "{1675E08B-7E62-4C5D-8FF6-E48AFD36442A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SeaSkyNavigator", "WeatherApp\SeaSkyNavigator.csproj", "{B7AA5BFD-8559-4A1E-A5A9-384C1B56BEF3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1675E08B-7E62-4C5D-8FF6-E48AFD36442A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1675E08B-7E62-4C5D-8FF6-E48AFD36442A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1675E08B-7E62-4C5D-8FF6-E48AFD36442A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1675E08B-7E62-4C5D-8FF6-E48AFD36442A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B7AA5BFD-8559-4A1E-A5A9-384C1B56BEF3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B7AA5BFD-8559-4A1E-A5A9-384C1B56BEF3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B7AA5BFD-8559-4A1E-A5A9-384C1B56BEF3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B7AA5BFD-8559-4A1E-A5A9-384C1B56BEF3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {3E502926-D63F-4086-B265-BD482C758E38}
+		SolutionGuid = {E7E30610-E8AF-4379-9672-888CD2C75C1D}
 	EndGlobalSection
 EndGlobal
