+++ WeatherApp/SeaSkyNavigator.http
@@ -1,0 +1,6 @@
+@SeaSkyNavigator_HostAddress = http://localhost:5091
+
+GET {{SeaSkyNavigator_HostAddress}}/weatherforecast/
+Accept: application/json
+
+###
